+++ demisto_sdk/__main__.py
@@ -3364,6 +3364,13 @@ def update_content_graph(
     default=False,
 )
 @click.option(
+    "-ns",
+    "--no-staged",
+    help="Whether to skip collecting staged files",
+    is_flag=True,
+    default=False,
+)
+@click.option(
     "-g",
     "--git-diff",
     help="Whether to use git to determine which files to run on",
@@ -3442,6 +3449,7 @@ def pre_commit(
     ctx,
     input: str,
     staged_only: bool,
+    no_staged: bool,
     git_diff: bool,
     all_files: bool,
     mode: Optional[str],
@@ -3476,6 +3484,7 @@ def pre_commit(
         pre_commit_manager(
             input_files,
             staged_only,
+            no_staged,
             git_diff,
             all_files,
             mode,